carlitos/browser-use/requirements.txt
browser-use>=0.1.45
langchain-openai>=0.0.5
python-dotenv>=1.0.0
patchright>=0.1.0 
playwright>=1.45.0